LayoutTests/svg/animations/animateTransform-pattern-transform-expected.txt

SVG 1.1 dynamic animation tests

Tests animation of 'patternTransform'. Should result in a 100x100 rect and only PASS messages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pattern.patternTransform.animVal.numberOfItems is 0
PASS pattern.patternTransform.baseVal.numberOfItems is 0
PASS pattern.patternTransform.animVal.numberOfItems is 1
PASS pattern.patternTransform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_SCALE
PASS pattern.patternTransform.animVal.getItem(0).matrix.a is 1
PASS pattern.patternTransform.baseVal.numberOfItems is 0
PASS pattern.patternTransform.animVal.numberOfItems is 1
PASS pattern.patternTransform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_SCALE
PASS pattern.patternTransform.animVal.getItem(0).matrix.a is 1.5
PASS pattern.patternTransform.baseVal.numberOfItems is 0
PASS pattern.patternTransform.animVal.numberOfItems is 1
PASS pattern.patternTransform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_SCALE
PASS pattern.patternTransform.animVal.getItem(0).matrix.a is 2
PASS pattern.patternTransform.baseVal.numberOfItems is 0
PASS pattern.patternTransform.animVal.numberOfItems is 0
PASS pattern.patternTransform.baseVal.numberOfItems is 0
PASS successfullyParsed is true

TEST COMPLETE
